+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.5.3] - Unreleased
+## [0.6.0] - Unreleased
+
+### Changed
+
+- Renamed the project to `rocketry`
+- The `-v --verbose` option has been renamed to `-d --debug`
 
 ### Fixed